+++ src/methods/query/01-time.js
@@ -163,17 +163,24 @@ const methods = {
     return 'night'
   },
 
-  //parse a proper iso string
+  //get/set a traditional iso string
   iso: function (num) {
     if (num !== undefined) {
       return this.set(num)
     }
     return this.format('iso')
   },
+  // get/set a new iso string
+  isoFull: function (num) {
+    if (num !== undefined) {
+      return this.set(num)
+    }
+    return this.format('iso-full')
+  },
   epochSeconds: function (num) {
     if (num !== undefined) {
       this.epoch = num * 1000
-      return
+      return this
     }
     return Math.floor(this.epoch / 1000)
   }